Migrate skill_fetcher tests and update references
