+++ core/controllers/skill_mastery.py
@@ -20,6 +20,7 @@ from __future__ import unicode_literals  # pylint: disable=import-only-modules
 from core.controllers import acl_decorators
 from core.controllers import base
 from core.domain import skill_domain
+from core.domain import skill_fetchers
 from core.domain import skill_services
 import feconf
 import utils
@@ -51,7 +52,7 @@ class SkillMasteryDataHandler(base.BaseHandler):
             raise self.InvalidInputException('Invalid skill ID %s' % skill_id)
 
         try:
-            skill_services.get_multi_skills(skill_ids)
+            skill_fetchers.get_multi_skills(skill_ids)
         except Exception as e:
             raise self.PageNotFoundException(e)
 
@@ -117,7 +118,7 @@ class SkillMasteryDataHandler(base.BaseHandler):
                 new_degrees_of_mastery[skill_id] = 1.0
 
         try:
-            skill_services.get_multi_skills(skill_ids)
+            skill_fetchers.get_multi_skills(skill_ids)
         except Exception as e:
             raise self.PageNotFoundException(e)
