+++ core/controllers/story_viewer.py
@@ -21,7 +21,7 @@ from constants import constants
 from core.controllers import acl_decorators
 from core.controllers import base
 from core.domain import question_services
-from core.domain import skill_services
+from core.domain import skill_fetchers
 from core.domain import story_fetchers
 from core.domain import story_services
 from core.domain import summary_services
@@ -136,7 +136,7 @@ class StoryProgressHandler(base.BaseHandler):
 
         # If there are no questions for any of the acquired skills that the
         # learner has completed, do not show review tests.
-        acquired_skills = skill_services.get_multi_skills(
+        acquired_skills = skill_fetchers.get_multi_skills(
             story.get_acquired_skill_ids_for_node_ids(
                 completed_node_ids
             ))
