+++ core/controllers/topics_and_skills_dashboard.py
@@ -368,11 +368,11 @@ class MergeSkillHandler(base.BaseHandler):
         """Handles the POST request."""
         old_skill_id = self.payload.get('old_skill_id')
         new_skill_id = self.payload.get('new_skill_id')
-        new_skill = skill_services.get_skill_by_id(new_skill_id, strict=False)
+        new_skill = skill_fetchers.get_skill_by_id(new_skill_id, strict=False)
         if new_skill is None:
             raise self.PageNotFoundException(
                 Exception('The new skill with the given id doesn\'t exist.'))
-        old_skill = skill_services.get_skill_by_id(old_skill_id, strict=False)
+        old_skill = skill_fetchers.get_skill_by_id(old_skill_id, strict=False)
         if old_skill is None:
             raise self.PageNotFoundException(
                 Exception('The old skill with the given id doesn\'t exist.'))
