+++ core/domain/skill_jobs_one_off_test.py
@@ -24,6 +24,7 @@ import ast
 from constants import constants
 from core.domain import skill_domain
 from core.domain import skill_jobs_one_off
+from core.domain import skill_fetchers
 from core.domain import skill_services
 from core.platform import models
 from core.tests import test_utils
@@ -81,7 +82,7 @@ class SkillMigrationOneOffJobTests(test_utils.GenericTestBase):
 
         # Verify the skill is exactly the same after migration.
         updated_skill = (
-            skill_services.get_skill_by_id(self.SKILL_ID))
+            skill_fetchers.get_skill_by_id(self.SKILL_ID))
         self.assertEqual(
             updated_skill.skill_contents_schema_version,
             feconf.CURRENT_SKILL_CONTENTS_SCHEMA_VERSION)
@@ -111,7 +112,7 @@ class SkillMigrationOneOffJobTests(test_utils.GenericTestBase):
 
         # Ensure the skill is deleted.
         with self.assertRaisesRegexp(Exception, 'Entity .* not found'):
-            skill_services.get_skill_by_id(self.SKILL_ID)
+            skill_fetchers.get_skill_by_id(self.SKILL_ID)
 
         # Start migration job on sample skill.
         job_id = (
@@ -124,7 +125,7 @@ class SkillMigrationOneOffJobTests(test_utils.GenericTestBase):
 
         # Ensure the skill is still deleted.
         with self.assertRaisesRegexp(Exception, 'Entity .* not found'):
-            skill_services.get_skill_by_id(self.SKILL_ID)
+            skill_fetchers.get_skill_by_id(self.SKILL_ID)
 
         output = skill_jobs_one_off.SkillMigrationOneOffJob.get_output(job_id)
         expected = [[u'skill_deleted',
@@ -180,7 +181,7 @@ class SkillMigrationOneOffJobTests(test_utils.GenericTestBase):
 
         # Verify that the skill migrates correctly.
         updated_skill = (
-            skill_services.get_skill_by_id(self.SKILL_ID))
+            skill_fetchers.get_skill_by_id(self.SKILL_ID))
 
         self.assertEqual(
             updated_skill.skill_contents_schema_version,
